refactor: rename sortCriteria.ts into sort-criteria.ts
