+++ src/lib/components/composites/select/SortOptionsSelect.svelte
@@ -5,7 +5,7 @@
         SortCriteria,
         SortDirection,
         type SortOptionLabel,
-    } from "$lib/model/sortCriteria";
+    } from "$lib/model/sort-criteria";
     import {
         ArrowDown,
         ArrowDown10,
