+++ tests/unit/utils/sorters.test.ts
@@ -2,7 +2,7 @@ import { describe, expect, test as baseTest } from "vitest";
 import { ProjectPapers } from "$tests/example-data";
 import { type Project_Paper } from "$lib/model/api/project";
 import { sortProjectPaper } from "$lib/utils/sorters";
-import { SortCriteria, SortDirection } from "$lib/model/sortCriteria";
+import { SortCriteria, SortDirection } from "$lib/model/sort-criteria";
 import { createPaper, createProjectPaper } from "$tests/model-builder";
 
 const test = baseTest.extend<{
